프로그래머스 두 원 사이의 정수 쌍 진행 중

+++ programmers/src/main/java/levelTwo/Dotting.java
@@ -3,12 +3,12 @@ package levelTwo;
 public class Dotting {
 
     public long solution(int k, int d) {
-        int answer = 0;
+        long answer = 0;
         long dSquared = (long) d * d;
         for (int i = 0; i <= d; i += k) {
             long maxYSquared = dSquared - (long) i * i;
-            int maxY = (int) Math.sqrt(maxYSquared);
-            answer += Math.ceil(maxY / k)+1;
+            long maxY = (long) Math.sqrt(maxYSquared);
+            answer += (long) (Math.ceil((double) maxY / k) + 1);
         }
         return answer;
     }
